mrc-weight-update-export(wu-move-from-disk-move,mechanical_provable): Move update_weights_from_disk onto WeightUpdater (cut+paste)
