+++ python/sglang/srt/speculative/eagle_worker_v2.py
@@ -1738,13 +1738,12 @@ class EAGLEWorkerV2(BaseSpecWorker):
         return out
 
     def update_weights_from_disk(self, recv_req: UpdateWeightFromDiskReqInput):
-        from sglang.srt.model_executor.model_runner import ModelRunner
-
-        success, message = ModelRunner.update_weights_from_disk(
-            self._draft_worker.draft_runner.weight_updater,
-            recv_req.model_path,
-            recv_req.load_format,
-            recapture_cuda_graph=recv_req.recapture_cuda_graph,
+        success, message = (
+            self._draft_worker.draft_runner.weight_updater.update_weights_from_disk(
+                recv_req.model_path,
+                recv_req.load_format,
+                recapture_cuda_graph=recv_req.recapture_cuda_graph,
+            )
         )
         if not success:
             return success, message
